requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.13.2; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
async-timeout==5.0.1; python_version >= '3.8'
attrs==25.4.0; python_version >= '3.9'
discord==2.3.2
discord.py==2.6.4; python_version >= '3.8'
fastapi==0.111.0
frozenlist==1.8.0; python_version >= '3.9'
httpx==0.28.1
idna==3.11; python_version >= '3.8'
itsdangerous==2.2.0
Jinja2==3.1.2
multidict==6.7.0; python_version >= '3.9'
peewee==3.18.3
PyMySQL==1.1.2
propcache==0.4.1; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
pyyaml==6.0.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
yarl==1.22.0; python_version >= '3.9'
